alibaba ID type

+++ API/AuthenticationAPI/Models/AlibabaDto.cs
@@ -61,8 +61,8 @@
 		public string responseId { get; set; }
 
 		[Required]
-		[Range(1, Int32.MaxValue, ErrorMessage = "AliMemberId is invalid")]
-		public int aliMemberId { get; set; }
+		[Range(1, Int64.MaxValue, ErrorMessage = "AliMemberId is invalid")]
+		public decimal aliMemberId { get; set; }
 
 		[Required(AllowEmptyStrings = false)]
 		[Range(1, Int32.MaxValue, ErrorMessage = "ezbob customer ID is invalid")]
